requirements.txt
aiohttp
arrow
asttokens
asyncio
boto3
botocore
certifi
debugpy
elastic-transport
elasticsearch
fastjsonschema
json5
jsonpointer
jsonschema
jsonschema-specifications
logstash
logstash_formatter
pylogbeat
python-dateutil
python-json-logger
python-logstash-async
PyYAML
